+++ src/wenet_survey/settings.py
@@ -9,34 +9,56 @@ https://docs.djangoproject.com/en/3.2/topics/settings/
 For the full list of settings and their values, see
 https://docs.djangoproject.com/en/3.2/ref/settings/
 """
-
+import logging.config
+import os
 from pathlib import Path
+
+from common.log.logging import get_logging_configuration
+
+logging.config.dictConfig(get_logging_configuration("wenet-survey"))
+
+logger = logging.getLogger(__name__)
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
+# Environment variables
+REQUIRED_ENV_VARS = ["WENET_APP_ID", "WENET_APP_SECRET", "WENET_INSTANCE_URL", "OAUTH_CALLBACK_URL"]
+
+for env_var in REQUIRED_ENV_VARS:
+    if os.getenv(env_var, None) is None:
+        raise ValueError(f"Missing required environment variable: [{env_var}]")
+
+WENET_APP_ID = os.getenv("WENET_APP_ID")
+WENET_APP_SECRET = os.getenv("WENET_APP_SECRET")
+WENET_INSTANCE_URL = os.getenv("WENET_INSTANCE_URL")
+OAUTH_CALLBACK_URL = os.getenv("OAUTH_CALLBACK_URL")
+BASE_URL = os.getenv("BASE_URL", "")
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/3.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'django-insecure-8jp0rb79f((j*#2604yhh5it&im25jni8@&t136ccnyb02yi_c'
+SECRET_KEY = os.getenv("SECRET_KEY") if os.getenv("SECRET_KEY", None) else 'django-insecure-8jp0rb79f((j*#2604yhh5it&im25jni8@&t136ccnyb02yi_c'
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
+DEBUG = os.getenv("DEBUG", "TRUE").upper() == "TRUE"
 
-ALLOWED_HOSTS = []
+ALLOWED_HOSTS = os.getenv("ALLOWED_HOSTS").split(";") if os.getenv("ALLOWED_HOSTS", None) is not None else []
 
 
 # Application definition
 
 INSTALLED_APPS = [
+    'authentication.apps.AuthenticationConfig',
+    'survey.apps.SurveyConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'rest_framework',
 ]
 
 MIDDLEWARE = [