+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
 ## 0.*
+
+### next
+
+* Set up the project with basic HTML pages.
+* Added support to WeNet OAuth2 authentication flow.